+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/CardField.cs
@@ -1,0 +1,29 @@
+﻿namespace TrelloDotNet.AutomationEngine.Model
+{
+    /// <summary>
+    /// Represents a Field on a Trello Card
+    /// </summary>
+    public enum CardField
+    {
+        /// <summary>
+        /// The Name Field
+        /// </summary>
+        Name,
+        /// <summary>
+        /// The Description Field
+        /// </summary>
+        Description,
+        /// <summary>
+        /// The Start Date Field
+        /// </summary>
+        Start,
+        /// <summary>
+        /// The Due Date Field
+        /// </summary>
+        Due,
+        /// <summary>
+        /// The Due Complete 
+        /// </summary>
+        DueComplete
+    }
+}