+++ Changelog.md
@@ -7,6 +7,7 @@
 - Added: Struct WebhookActionTypes that list all Type of Webhook events
 - Added: Method 'DeleteChecklistAsync'
 - Fixed: Various incorrect XML Summaries and parameter names (all non breaking changes)
+- Fixed: Null values in strings for update of objects are now considered empty string so you do not end up with 'null' values (aka the word 'null' as a string)
 
 
 ## 1.2.1 (25th of Feb. 2023)